+++ projects/np-ui-lib/src/lib/np-list/np-list.module.ts
@@ -2,12 +2,14 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpListComponent } from './np-list.component';
 import { NpUtilityModule } from '../np-utility/np-utility.module';
+import { NpPaginatorModule } from '../np-paginator/np-paginator.module';
 
 @NgModule({
   declarations: [NpListComponent],
   imports: [
     CommonModule,
-    NpUtilityModule
+    NpUtilityModule,
+    NpPaginatorModule
   ],
   exports: [NpListComponent]
 })